move rich text color picker logic from MarkdownEditor to SpectrumMDE
