+++ packages/bbui/src/Markdown/MarkdownEditor.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
   import type EasyMDE from "easymde"
-  import { onDestroy, tick } from "svelte"
+  import { onDestroy } from "svelte"
   import SpectrumMDE from "./SpectrumMDE.svelte"
-  import ColorPicker from "../ColorPicker/ColorPicker.svelte"
   import { createEventDispatcher } from "svelte"
 
   export let value: string | null = null
@@ -22,191 +21,24 @@
     value: (_value?: string) => string
   }
   let mde: EditorInstance | null = null
-  const colorDefaults = {
-    text: "#d73f09",
-    highlight: "#fff59d",
-  } as const
-  const modeConfig = {
-    text: {
-      toolbarButtonName: "text-color",
-      wrapperTag: "span",
-      stylePrefix: "color",
-    },
-    highlight: {
-      toolbarButtonName: "text-highlight",
-      wrapperTag: "mark",
-      stylePrefix: "background-color",
-    },
-  } as const
-  type ColorMode = keyof typeof colorDefaults
-  interface EditorSelectionRange {
-    anchor: { line: number; ch: number }
-    head: { line: number; ch: number }
-  }
-  interface EasyMDEWithToolbar extends EasyMDE {
-    toolbarElements?: Record<string, HTMLElement>
-  }
-  let colorPickerAnchor: HTMLDivElement | undefined = undefined
-  let colorPickerX = 0
-  let colorPickerY = 0
-  let colorPickerKey = 0
-  let activeMode: ColorMode = "text"
-  let selectedColors: Record<ColorMode, string> = {
-    text: colorDefaults.text,
-    highlight: colorDefaults.highlight,
-  }
-  let pendingSelections: EditorSelectionRange[] | null = null
-  let lastNonEmptySelections: EditorSelectionRange[] | null = null
-  let eventsBoundTo: EditorInstance | null = null
+  let blurBoundTo: EditorInstance | null = null
 
-  const cloneSelections = (selections: EditorSelectionRange[]) =>
-    selections.map(selection => ({
-      anchor: { ...selection.anchor },
-      head: { ...selection.head },
-    }))
-
-  const hasSelectedText = (editor: EasyMDE) =>
-    editor.codemirror.getSelections().some((text: string) => text.length > 0)
-
-  const getSelections = (editor: EasyMDE) =>
-    cloneSelections(
-      editor.codemirror.listSelections() as EditorSelectionRange[]
-    )
-
-  const cacheSelection = () => {
-    if (!mde) {
+  const bindBlurHandler = (editor: EditorInstance) => {
+    if (blurBoundTo === editor) {
       return
     }
-    if (!hasSelectedText(mde)) {
-      return
-    }
-    lastNonEmptySelections = getSelections(mde)
-  }
-
-  const bindEditorEvents = (editor: EditorInstance) => {
-    if (eventsBoundTo === editor) {
-      return
-    }
-    if (eventsBoundTo) {
-      eventsBoundTo.codemirror.off("blur", update)
-      eventsBoundTo.codemirror.off("cursorActivity", cacheSelection)
+    if (blurBoundTo) {
+      blurBoundTo.codemirror.off("blur", update)
     }
     editor.codemirror.on("blur", update)
-    editor.codemirror.on("cursorActivity", cacheSelection)
-    eventsBoundTo = editor
+    blurBoundTo = editor
   }
-
-  const openColorPicker = async (editor: EasyMDE, mode: ColorMode) => {
-    const currentSelections = getSelections(editor)
-    const activeSelections = hasSelectedText(editor)
-      ? currentSelections
-      : lastNonEmptySelections || currentSelections
-    pendingSelections = cloneSelections(activeSelections)
-    activeMode = mode
-    const toolbarButtonName = modeConfig[mode].toolbarButtonName
-    const toolbarButton = (editor as EasyMDEWithToolbar).toolbarElements?.[
-      toolbarButtonName
-    ]
-    if (toolbarButton) {
-      const rect = toolbarButton.getBoundingClientRect()
-      colorPickerX = Math.round(rect.left + rect.width / 2)
-      colorPickerY = Math.round(rect.bottom)
-    }
-    colorPickerKey += 1
-    await tick()
-    const trigger = colorPickerAnchor?.querySelector(".preview") as
-      | HTMLElement
-      | undefined
-    trigger?.click()
-  }
-
-  const applyStyledSelections = (color: string, mode: ColorMode) => {
-    if (!pendingSelections || !mde) {
-      return
-    }
-    const safeColor = color.replace(/[<>"']/g, "")
-    if (!safeColor) {
-      return
-    }
-    mde.codemirror.focus()
-    mde.codemirror.setSelections(pendingSelections)
-    const selectedTexts = mde.codemirror.getSelections()
-    const hasText = selectedTexts?.some((text: string) => text.length > 0)
-    if (!selectedTexts?.length || !hasText) {
-      pendingSelections = null
-      return
-    }
-    const { wrapperTag, stylePrefix } = modeConfig[mode]
-    const styleAttr = `${stylePrefix}: ${safeColor};`
-    const replacements = selectedTexts.map(
-      (text: string) =>
-        `<${wrapperTag} style="${styleAttr}">${text}</${wrapperTag}>`
-    )
-    mde.codemirror.replaceSelections(replacements)
-    pendingSelections = null
-  }
-
-  const onColorChange = (event: CustomEvent<string | undefined>) => {
-    const color = event.detail?.trim()
-    if (!color) {
-      return
-    }
-    selectedColors = {
-      ...selectedColors,
-      [activeMode]: color,
-    }
-    applyStyledSelections(color, activeMode)
-  }
-
-  onDestroy(() => {
-    if (!eventsBoundTo) {
-      return
-    }
-    eventsBoundTo.codemirror.off("blur", update)
-    eventsBoundTo.codemirror.off("cursorActivity", cacheSelection)
-  })
-
-  const textColorToolbarButton = {
-    name: "text-color",
-    action: (editor: EasyMDE) => openColorPicker(editor, "text"),
-    className: "fa fa-font",
-    title: "Text Color",
-  }
-
-  const highlightToolbarButton = {
-    name: "text-highlight",
-    action: (editor: EasyMDE) => openColorPicker(editor, "highlight"),
-    className: "fa fa-paint-brush",
-    title: "Highlight Color",
-  }
-
-  const defaultToolbar = [
-    "bold",
-    "italic",
-    "heading",
-    "|",
-    "quote",
-    "unordered-list",
-    "ordered-list",
-    "|",
-    "link",
-    "image",
-    "|",
-    textColorToolbarButton,
-    highlightToolbarButton,
-    "|",
-    "preview",
-    "side-by-side",
-    "fullscreen",
-    "|",
-    "guide",
-  ]
 
   // Ensure the value is updated if the value prop changes outside the editor's
   // control
   $: checkValue(mde, value)
   $: if (mde) {
-    bindEditorEvents(mde)
+    bindBlurHandler(mde)
   }
   $: if ((readonly || disabled) && mde) {
     mde.togglePreview?.()
@@ -226,16 +58,12 @@
     dispatch("change", latestValue)
   }
 
-  const getToolbar = (
-    disabled: boolean,
-    readonly: boolean,
-    easyMDEOptions: Record<string, any>
-  ) => {
-    if (disabled || readonly) {
-      return false
+  onDestroy(() => {
+    if (blurBoundTo) {
+      blurBoundTo.codemirror.off("blur", update)
     }
-    return easyMDEOptions.toolbar ?? defaultToolbar
-  }
+  })
+
 </script>
 
 {#key height}
@@ -250,32 +78,7 @@
       initialValue: value,
       placeholder,
       ...easyMDEOptions,
-      toolbar: getToolbar(disabled, readonly, easyMDEOptions),
+      toolbar: disabled || readonly ? false : easyMDEOptions?.toolbar,
     }}
   />
 {/key}
-
-<div
-  bind:this={colorPickerAnchor}
-  class="budibase-color-picker-anchor"
-  style={`left:${colorPickerX}px;top:${colorPickerY}px;`}
->
-  {#key colorPickerKey}
-    <ColorPicker
-      value={selectedColors[activeMode]}
-      size="S"
-      on:change={onColorChange}
-    />
-  {/key}
-</div>
-
-<style>
-  .budibase-color-picker-anchor {
-    position: fixed;
-    width: 1px;
-    height: 1px;
-    overflow: hidden;
-    opacity: 0;
-    pointer-events: none;
-  }
-</style>
