+++ packages/bbui/src/Markdown/SpectrumMDE.svelte
@@ -1,17 +1,192 @@
 <script lang="ts">
   import type EasyMDE from "easymde"
   import "easymde/dist/easymde.min.css"
-  import { onDestroy, onMount } from "svelte"
+  import { onDestroy, onMount, tick } from "svelte"
+  import ColorPicker from "../ColorPicker/ColorPicker.svelte"
 
   export let height: string | null = null
   export let scroll: boolean = true
   export let easyMDEOptions: Record<string, any> | null = null
-  export let mde: EasyMDE | null = null
+  interface EditorInstance extends EasyMDE {
+    toolbarElements?: Record<string, HTMLElement>
+  }
+  interface EditorSelectionRange {
+    anchor: { line: number; ch: number }
+    head: { line: number; ch: number }
+  }
+  type ColorMode = "text" | "highlight"
+
+  export let mde: EditorInstance | null = null
   export let id: string | null = null
   export let fullScreenOffset: { x: string; y: string } | null = null
   export let disabled: boolean = false
 
   let element: HTMLTextAreaElement | undefined = undefined
+  let colorPickerAnchor: HTMLDivElement | undefined = undefined
+  let colorPickerX = 0
+  let colorPickerY = 0
+  let colorPickerKey = 0
+  let activeMode: ColorMode = "text"
+  let selectedColors: Record<ColorMode, string> = {
+    text: "#d73f09",
+    highlight: "#fff59d",
+  }
+  let pendingSelections: EditorSelectionRange[] | null = null
+  let lastNonEmptySelections: EditorSelectionRange[] | null = null
+  let cursorBoundTo: EditorInstance | null = null
+
+  const modeConfig = {
+    text: {
+      toolbarButtonName: "text-color",
+      wrapperTag: "span",
+      stylePrefix: "color",
+      className: "fa fa-font",
+      title: "Text Color",
+    },
+    highlight: {
+      toolbarButtonName: "text-highlight",
+      wrapperTag: "mark",
+      stylePrefix: "background-color",
+      className: "fa fa-paint-brush",
+      title: "Highlight Color",
+    },
+  } as const
+
+  const cloneSelections = (selections: EditorSelectionRange[]) =>
+    selections.map(selection => ({
+      anchor: { ...selection.anchor },
+      head: { ...selection.head },
+    }))
+
+  const hasSelectedText = (editor: EditorInstance) =>
+    editor.codemirror.getSelections().some((text: string) => text.length > 0)
+
+  const getSelections = (editor: EditorInstance) =>
+    cloneSelections(
+      editor.codemirror.listSelections() as EditorSelectionRange[]
+    )
+
+  const cacheSelection = () => {
+    if (!mde || !hasSelectedText(mde)) {
+      return
+    }
+    lastNonEmptySelections = getSelections(mde)
+  }
+
+  const bindCursorListener = (editor: EditorInstance) => {
+    if (cursorBoundTo === editor) {
+      return
+    }
+    if (cursorBoundTo) {
+      cursorBoundTo.codemirror.off("cursorActivity", cacheSelection)
+    }
+    editor.codemirror.on("cursorActivity", cacheSelection)
+    cursorBoundTo = editor
+  }
+
+  const openColorPicker = async (editor: EasyMDE, mode: ColorMode) => {
+    const typedEditor = editor as EditorInstance
+    const currentSelections = getSelections(typedEditor)
+    const activeSelections = hasSelectedText(typedEditor)
+      ? currentSelections
+      : lastNonEmptySelections || currentSelections
+    pendingSelections = cloneSelections(activeSelections)
+    activeMode = mode
+    const toolbarButtonName = modeConfig[mode].toolbarButtonName
+    const toolbarButton = typedEditor.toolbarElements?.[toolbarButtonName]
+    if (toolbarButton) {
+      const rect = toolbarButton.getBoundingClientRect()
+      colorPickerX = Math.round(rect.left + rect.width / 2)
+      colorPickerY = Math.round(rect.bottom)
+    }
+    colorPickerKey += 1
+    await tick()
+    const trigger = colorPickerAnchor?.querySelector(".preview") as
+      | HTMLElement
+      | undefined
+    trigger?.click()
+  }
+
+  const applyStyledSelections = (color: string, mode: ColorMode) => {
+    if (!pendingSelections || !mde) {
+      return
+    }
+    const safeColor = color.replace(/[<>"']/g, "")
+    if (!safeColor) {
+      return
+    }
+    mde.codemirror.focus()
+    mde.codemirror.setSelections(pendingSelections)
+    const selectedTexts = mde.codemirror.getSelections()
+    const hasText = selectedTexts?.some((text: string) => text.length > 0)
+    if (!selectedTexts?.length || !hasText) {
+      pendingSelections = null
+      return
+    }
+    const { wrapperTag, stylePrefix } = modeConfig[mode]
+    const styleAttr = `${stylePrefix}: ${safeColor};`
+    const replacements = selectedTexts.map(
+      (text: string) =>
+        `<${wrapperTag} style="${styleAttr}">${text}</${wrapperTag}>`
+    )
+    mde.codemirror.replaceSelections(replacements)
+    pendingSelections = null
+  }
+
+  const onColorChange = (event: CustomEvent<string | undefined>) => {
+    const color = event.detail?.trim()
+    if (!color) {
+      return
+    }
+    selectedColors = {
+      ...selectedColors,
+      [activeMode]: color,
+    }
+    applyStyledSelections(color, activeMode)
+  }
+
+  const textColorToolbarButton = {
+    name: modeConfig.text.toolbarButtonName,
+    action: (editor: EasyMDE) => openColorPicker(editor, "text"),
+    className: modeConfig.text.className,
+    title: modeConfig.text.title,
+  }
+
+  const highlightToolbarButton = {
+    name: modeConfig.highlight.toolbarButtonName,
+    action: (editor: EasyMDE) => openColorPicker(editor, "highlight"),
+    className: modeConfig.highlight.className,
+    title: modeConfig.highlight.title,
+  }
+
+  const defaultToolbar = [
+    "bold",
+    "italic",
+    "heading",
+    "|",
+    "quote",
+    "unordered-list",
+    "ordered-list",
+    "|",
+    "link",
+    "image",
+    "|",
+    textColorToolbarButton,
+    highlightToolbarButton,
+    "|",
+    "preview",
+    "side-by-side",
+    "fullscreen",
+    "|",
+    "guide",
+  ]
+
+  const getToolbar = (options: Record<string, any> | null) => {
+    if (options?.toolbar !== undefined) {
+      return options.toolbar
+    }
+    return defaultToolbar
+  }
 
   onMount(async () => {
     height = height || "200px"
@@ -24,10 +199,15 @@
       maxHeight: scroll ? height : undefined,
       minHeight: scroll ? undefined : height,
       ...easyMDEOptions,
+      toolbar: getToolbar(easyMDEOptions),
     })
+    bindCursorListener(mde)
   })
 
   onDestroy(() => {
+    if (cursorBoundTo) {
+      cursorBoundTo.codemirror.off("cursorActivity", cacheSelection)
+    }
     mde?.toTextArea()
   })
 
@@ -45,7 +225,26 @@
   <textarea disabled {id} bind:this={element}></textarea>
 </div>
 
+<div
+  bind:this={colorPickerAnchor}
+  class="budibase-color-picker-anchor"
+  style={`left:${colorPickerX}px;top:${colorPickerY}px;`}
+>
+  {#key colorPickerKey}
+    <ColorPicker value={selectedColors[activeMode]} size="S" on:change={onColorChange} />
+  {/key}
+</div>
+
 <style>
+  .budibase-color-picker-anchor {
+    position: fixed;
+    width: 1px;
+    height: 1px;
+    overflow: hidden;
+    opacity: 0;
+    pointer-events: none;
+  }
+
   /* Disabled styles */
   .disabled :global(textarea) {
     display: none;
